Split `script` into multiple modules
