+++ src/pattern.rs
@@ -1,4 +1,6 @@
-use crate::script::{Control::*, LargeValue::*, Normal::*, SmallValue::*, *};
+use crate::opcode::{Control::*, LargeValue::*, Normal::*, SmallValue::*, *};
+use crate::script::*;
+use crate::scriptnum::*;
 
 // Much of this comes from
 // https://gist.github.com/str4d/9d80f1b60e6787310897044502cb025b
